refactor poll controller to component
